Fixed issue where AndroidManifest.xml was not getting updated. Moved to after platform add hook.
