+++ tools/postinstall-android.js
@@ -19,34 +19,6 @@ var copyFile = function(srcPath, targetPath) {
     fs.createReadStream(srcPath).pipe(fs.createWriteStream(targetPath));
 };
 
-var fixFile = function(path, fix) {
-    fs.writeFileSync(path, fix(fs.readFileSync(path, 'utf8')));
-};
-
-// Function to fix AndroidManifest.xml
-var fixAndroidManifest = function(data) {
-
-    // Fix application tag
-    var appName = "com.salesforce.androidsdk.phonegap.app.HybridApp";
-
-    // In case the script was run twice
-    if (data.indexOf(appName) == -1) {
-        var applicationTag = '<application android:hardwareAccelerated="true" android:icon="@drawable/sf__icon" android:label="@string/app_name" android:manageSpaceActivity="com.salesforce.androidsdk.ui.ManageSpaceActivity" android:name="' + appName + '">'
-        data = data.replace(/<application [^>]*>/, applicationTag);
-
-        // Comment out first activity
-        data = data.replace(/<activity/, "<!--<activity");
-        data = data.replace(/<\/activity>/, "</activity>-->");
-
-        // Change min sdk version
-        data = data.replace(/android\:minSdkVersion\=\"10\"/, 'android:minSdkVersion="19"');
-
-        // Change target api
-        data = data.replace(/android\:targetSdkVersion\=\"22\"/, 'android:targetSdkVersion="' + targetAndroidApi + '"');
-    }
-    return data;
-};
-
 var getAndroidSDKToolPath = function() {
     var androidHomeDir = process.env.ANDROID_HOME;
     if (typeof androidHomeDir !== 'string') {
@@ -76,9 +48,6 @@ if (androidExePath === null) {
 var pluginRoot = path.join('plugins', 'com.salesforce');
 var libProjectRoot = path.join('plugins', 'com.salesforce', 'src', 'android', 'libs');
 var appProjectRoot = path.join('platforms', 'android');
-
-console.log('Fixing application AndroidManifest.xml');
-fixFile(path.join('platforms', 'android', 'AndroidManifest.xml'), fixAndroidManifest);
 
 console.log('Moving Salesforce libraries to the correct location');
 shelljs.cp('-R', path.join(libProjectRoot, 'SalesforceSDK'), appProjectRoot);
